Following up on the latest schema changes.
The internal ids ('_tx_id' and '_exon_id') are never exposed to the
user anymore and only used for the SQL joins. All the functions that
were returning some internal ids to the user are now returning the
external ids instead (i.e. 'tx_id' instead of '_tx_id' and 'exon_id'
instead of '_exon_id').


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/GenomicFeatures@44018 bc3139a8-67e5-0310-9ffc-ced21a209358