Changes to the sorting and fields returned by transcriptsBy etc.
This simplifies what is returned by the *By category of functions so
that they are all returning data that pertains only to the ranges
requested.  Also sorting now has an implicit order to it.  Ordering is
done by exon rank where when grouping by transcripts, but by
chromosome location otherwise.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/GenomicFeatures@44879 bc3139a8-67e5-0310-9ffc-ced21a209358