Fix flaw in heuristic for inferring exon ranks in makeTxDbFromGRanges()
makeTxDbFromGRanges() can guess the exon ranks of a given transcript either

  (a) based on their position on the chromosome

  (b) or by looking at the suffixes of the exon ids (e.g. .1, .2, etc...)
      if any

Previously (i.e. in GenomicFeatures < 1.53.1) it was trying (b) first, and
would fall back on (a) if (b) failed.

Starting with GenomicFeatures 1.53.1, it does the opposite: it tries (a)
first, and uses (b) only for exons for which (a) failed.

This change addresses the problem that the suffixes of the exon ids cannot
be trusted to infer the ranks of exons located on the minus strand. See
https://github.com/Bioconductor/GenomicFeatures/issues/59 for an example.