Calls to functions defined in txdbmaker are now redirected there
The following GenomicFeatures functions are now implemented in the new
txdbmaker package:
- makeTxDb
- supportedUCSCtables
- browseUCSCtrack
- makeTxDbFromUCSC
- getChromInfoFromBiomart
- makeTxDbFromBiomart
- makeTxDbFromEnsembl
- makeTxDbFromGRanges
- makeTxDbFromGFF
- supportedUCSCFeatureDbTracks
- supportedUCSCFeatureDbTables
- UCSCFeatureDbTableSchema
- makeFeatureDbFromUCSC
- supportedMiRBaseBuildValues
- makePackageName
- makeTxDbPackage
- makeTxDbPackageFromUCSC
- makeFDbPackageFromUCSC
- makeTxDbPackageFromBiomart

Note that they are still temporarily defined in GenomicFeatures but now
they just call the corresponding function in txdbmaker. Since this is a
temporary redirect, the user also gets a warning that tells them to use
the fully qualified name (e.g. txdbmaker::makeTxDbFromUCSC()) to call
the function.